Move helper function to Helper.js
