+++ src/Helper.js
@@ -7,3 +7,13 @@ class Helper {
 
 
 }
+
+export const helper = (films) => {
+  const randomIndex = Math.floor(Math.random() * 8);
+   const result = {
+       crawl: films.results[randomIndex].opening_crawl,
+       title: films.results[randomIndex].title,
+       episode: films.results[randomIndex].episode_id
+   }
+   return result
+ }
